+++ cgbeacon2/constants/__init__.py
@@ -4,7 +4,6 @@ from .query_errors import (
     NO_MANDATORY_PARAMS,
     NO_SECONDARY_PARAMS,
     NO_POSITION_PARAMS,
-    NO_SV_END_PARAM,
     INVALID_COORDINATES,
     INVALID_COORD_RANGE,
     BUILD_MISMATCH,
@@ -16,7 +15,7 @@ from .oauth_errors import (
     EXPIRED_TOKEN_SIGNATURE,
     INVALID_TOKEN_AUTH,
     NO_GA4GH_USERDATA,
-    PASSPORTS_ERROR
+    PASSPORTS_ERROR,
 )
 from .request_errors import MISSING_TOKEN, WRONG_SCHEME
 from .response_objs import QUERY_PARAMS_API_V1